style: set styling for branding and main sections
The page layout is divided onto the branding section and main part containing the content and form. Changes in styling for branding include the adition of logo image, background image and custom font.
The main section styling considers headers for content text and labels, text inputs, submission button and fieldset for the form.
Displays of structural elements are set to flex in most cases. Paye layout also required more arbitrary spaces for making it look similar to the reference image.